+++ src/main/java/com/recetasAPD/recetasAPD/dtos/CalificacionResponse.java
@@ -16,7 +16,7 @@ public class CalificacionResponse {
     private Integer idCalificacion;
     private Usuario usuario;
     private Receta receta;
-    private Integer calificacion;
+    private Float calificacion;
     private String comentarios;
